moved the app methods into a new folder and connected it to the other express app using router
